Refactor files layout
